% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_dssat.R
\name{add_cultivar}
\alias{add_cultivar}
\title{Add new cultivar to the CUL table}
\usage{
add_cultivar(cul, ccode, cname, ecode, ppars, gpars)
}
\arguments{
\item{cul}{a data frame/tibble with cultivar data, as obtained with DSSAT::read_cul}

\item{ccode}{a length 1 character with the 6-character cultivar code that links to DSSAT file X}

\item{cname}{a length 1 character with the cultivar name}

\item{ecode}{a length 1 character with the 6-character ecotype code that links to DSSAT ECO file}

\item{ppars}{a length 5 numeric vector containing phenology parameters (VSEN, PPSEN, P1, P5, PHINT)}

\item{gpars}{a length 4 numeric vector containing growth parameters (GRNO, MXFIL, STMMX, SLAP1)}
}
\value{
a tibble; input table with the added cultivar as last row
}
\description{
Add new cultivar to the CUL table
}
