% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_data.R
\name{apply_function}
\alias{apply_function}
\title{Apply a Wrangling Function to a Data Frame}
\usage{
apply_function(var, fun, args, df)
}
\arguments{
\item{var}{Character. The target variable for the function (\code{x}).}

\item{fun}{Character string containing the function definition.}

\item{args}{List of additional arguments for the function.}

\item{df}{Data frame to which the function will be applied.}
}
\value{
A data frame with the function applied.
}
\description{
Applies a user-supplied function (as a character string) to a data frame,
with specified variable and additional arguments.
}
