% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_dataset.R
\name{apply_master_key}
\alias{apply_master_key}
\title{Propagate the Master Key Through a Dataset}
\usage{
apply_master_key(dataset, model_config)
}
\arguments{
\item{dataset}{The input dataset as a list of data frames.}

\item{model_config}{The configuration block for the input data model.}
}
\value{
The dataset with the master key propagated to all relevant tables.
}
\description{
A pre-processing step that intelligently links all tables in a dataset by ensuring the master key is present.
}
\details{
The function's behavior depends on the number of unique master keys found:
\itemize{
\item \strong{Single-Experiment:} For tables missing the key, it adds a column containing the single unique master key value.
\item \strong{Multi-Experiment:} For "shared" tables (e.g., \code{WEATHER}, \code{SOIL}), it performs a \code{cross_join} to link every
master key to the shared data.
}
}
