% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_metadata.R
\name{apply_transformations}
\alias{apply_transformations}
\title{Apply Post-Processing Transformations to a Mapped Dataset}
\usage{
apply_transformations(
  dataset,
  data_model = c("icasa", "dssat"),
  config_path = "./inst/extdata/datamodels.yaml"
)
}
\arguments{
\item{dataset}{The mapped dataset as a list of named data frames.}

\item{data_model}{The target data model, either "icasa" or "dssat".}

\item{config_path}{The path to the \code{datamodels.yaml} configuration file.}
}
\value{
A list of data frames structured for the target data model; for DSSAT, a nested list with one sub-list
for each experiment.
}
\description{
A top-level function that applies model-specific structuring and aggregation logic to a dataset after the initial
mapping from a source model is complete.
}
\details{
This function acts as a dispatcher, calling the appropriate structuring functions based on the specified \code{data_model}.
\itemize{
\item For \strong{\code{data_model = "icasa"}}, it calls \code{structure_icasa_mngt()} to aggregate detailed management event data
into standardized "regimes". It also  performs initial aggregations on field-level data.
\item For \strong{\code{data_model = "dssat"}}, it orchestrates a series of calls to other specialized functions
(\code{structure_dssat_mngt}, \code{structure_dssat_sol}, \code{structure_dssat_wth}) to enrich metadata, generate standard IDs,
split the dataset by experiment, and format the output for DSSAT files.
}
}
