% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_utils.R
\name{attr_to_column}
\alias{attr_to_column}
\title{Append data frame attributes to the first non-join columns so that they are not lost at joining}
\usage{
attr_to_column(df1, df2, attr_name)
}
\value{
a list containing the two data frames with the attributes appended to the first non-join columns
}
\description{
Append data frame attributes to the first non-join columns so that they are not lost at joining
}
