% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_data.R
\name{build_composite_data}
\alias{build_composite_data}
\title{Build a Composite Data Frame from a List}
\usage{
build_composite_data(df_list, groups, action = "coalesce")
}
\arguments{
\item{df_list}{A list of data frames to combine.}

\item{groups}{A character vector of column names to group by for summarization.}

\item{action}{The summary method: \code{"coalesce"} (default) to take the first non-NA value per group, or \code{"average"} to compute
the mean.}
}
\value{
A single, summarized data frame.
}
\description{
Combines a list of data frames and summarizes the data by specified groups, either by taking the first non-NA value (\code{coalesce})
or by averaging.
}
\examples{
df1 <- data.frame(id = "A", val1 = 1, val2 = NA)
df2 <- data.frame(id = "A", val1 = NA, val2 = 10)
df3 <- data.frame(id = "B", val1 = 5, val2 = 20)
df_list <- list(df1, df2, df3)

# Coalesce values by group (default action)
build_composite_data(df_list, groups = "id")

# Average values by group
build_composite_data(df_list, groups = "id", action = "average")

}
