% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_simulation_files.R
\name{build_simulation_files}
\alias{build_simulation_files}
\title{Build, parameterize, and write a DSSAT dataset}
\usage{
build_simulation_files(
  dataset,
  write = FALSE,
  sol_append = FALSE,
  write_in_dssat_dir = TRUE,
  path = getwd(),
  control_config = NULL
)
}
\arguments{
\item{dataset}{(list) A named list of raw DSSAT data containing formatted DSSAT data such as:
\code{"MANAGEMENT"}, \code{"SOIL"}, \code{"WEATHER"}, \code{"SUMMARY"}, and \code{"TIME_SERIES"} elements.}

\item{write}{(logical) If \code{TRUE}, formatted files are written to disk. Default: \code{FALSE}.}

\item{sol_append}{(logical) If \code{TRUE}, soil profile data is appended as a new record at writing, provided a DSSAT soil file
with the same indentifier (institution) already exists at the write location}

\item{write_in_dssat_dir}{(logical) If \code{TRUE}, files are written to standard DSSAT subdirectories.
If \code{FALSE}, all files are written to the \code{path} directory. Default: \code{TRUE}.}

\item{path}{(character) Target directory for writing files. \strong{Only used if \code{write_in_dssat_dir = FALSE}}. Default: \code{getwd()}.}

\item{control_config}{(character) Path to a JSON/YAML configuration file specifying DSSAT simulation control parameters.
The file should contain key-value pairs where keys are parameter names (e.g., \code{"NITROGEN"}, \code{"WATER"}, \code{"RSEED"}, \code{"SDATE"})
and values are the desired settings. These values will override DSSAT's default simulation controls.
To view default values, run \code{data(dssat_default_simcontrols)}.}
}
\value{
A named list containing a simulation-ready, formatted, and parameterized DSSAT input data.
The list names will be \code{"MANAGEMENT"}, \code{"SOIL"}, \code{"WEATHER"}, \code{"SUMMARY"}, and \code{"TIME_SERIES"}.
}
\description{
Prepares a DSSAT dataset for a model run:
\enumerate{
\item Formats all raw data sections as proper inputs for DSSAT parser.
\item Writes provided simulation control parameters to MANAGEMENT file.
\item Optionally writes all files to their correct locations in DSSAT subdirectories or a user-specified location.
}
}
\details{
This function is a high-level wrapper that calls several internal helpers in sequence:
\enumerate{
\item \code{build_dssat_dataset} formats all raw data sections.
\item \code{set_dssat_controls} modifies the \code{MANAGEMENT} section with parameters from \code{control_args} config file.
\item If \code{write = TRUE}, \code{prepare_dssat_paths} resolves all output file paths
\item if \code{write_dssat_dataset = TRUE} DSSAT input files are written to the adequate location in the DSSAT directory;
if \code{write_dssat_dataset = FALSE}, files are written to the user-specified \code{path}.
}
}
\examples{
\dontrun{
# Assuming 'my_raw_dataset' is a list of raw data

# Just format and parameterize (here, using config file), without writing
formatted_data <- build_simulation_files(
  my_raw_dataset,
  control_config = "path/to/sim_controls.yaml"
)

# Format, parameterize, and write to a custom local directory
build_simulation_files(
  my_raw_dataset,
  write = TRUE,
  write_in_dssat_dir = FALSE,
  path = "./model_run_1",
  sol_append = FALSE,
  control_config = "path/to/sim_controls.json"
)

# Format, parameterize, and write to standard DSSAT directories
build_simulation_files(
  my_raw_dataset,
  control_config = "path/to/sim_controls.yaml",
  write = TRUE,
  write_in_dssat_dir = TRUE
)
}
}
