% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_initial_layers.R
\name{calculate_initial_layers}
\alias{calculate_initial_layers}
\title{Wrapper to calculate initial soil conditions for DSSAT simulations}
\usage{
calculate_initial_layers(
  soil_profile,
  percent_available_water = NULL,
  total_n_kgha = NULL
)
}
\arguments{
\item{soil_profile}{A data.frame representing a DSSAT soil profile.}

\item{percent_available_water}{A numeric value (0-100); available water for entire profile (\%).}

\item{total_n_kgha}{A numeric value; total nitrogen fo entire profile (kg/ha).}
}
\value{
A data.frame representing the DSSAT initial condition layers section ('SH2O', 'SNO3', 'SNH4')
for each soil layer defined in the input soil profile.
}
\description{
Calculates initial soil conditions for water and/or nitrogen based on single input values.
}
\examples{
# Define a sample soil profile
soil_df <- data.frame(
  SLB = c(15, 30, 60, 90),       # Depth at bottom of layer (cm)
  SLLL = c(0.10, 0.12, 0.15, 0.18), # Wilting Point (cm3/cm3)
  SDUL = c(0.25, 0.28, 0.32, 0.35), # Field Capacity (cm3/cm3)
  SBDM = c(1.5, 1.4, 1.4, 1.3)       # Bulk Density (g/cm3)
)

# Calculate both water and nitrogen
calculate_initial_conditions(soil_df, percent_available_water = 80, total_n_kgha = 25)

# Calculate only water
calculate_initial_conditions(soil_df, percent_available_water = 50)

# Calculate only nitrogen
calculate_initial_conditions(soil_df, total_n_kgha = 40)

}
