% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_data.R
\name{convert_unit}
\alias{convert_unit}
\title{Convert numeric variables based on units provided as strings}
\usage{
convert_unit(x, u1, u2)
}
\arguments{
\item{x}{a numeric to be converted}

\item{u1}{the input unit, in which x is expressed}

\item{u2}{the target unit, in which the output is expressed}
}
\value{
a converted numeric
}
\description{
####str unit formats
}
