% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_exp_design.R
\name{count_per_year}
\alias{count_per_year}
\title{Count Unique Values per Year}
\usage{
count_per_year(col, df, yr = "Year")
}
\arguments{
\item{col}{The name of the column to count unique values in (as a string).}

\item{df}{A data frame containing the data.}

\item{yr}{The name of the year column (as a string). Default is \code{"Year"}.}
}
\value{
A named integer vector giving the number of unique values in \code{col} for each year.
}
\description{
Counts the number of unique values in a specified column for each year in a data frame.
}
\examples{
df <- data.frame(
  Year = c(2000, 2000, 2001, 2001, 2001),
  Plot = c(1, 2, 1, 2, 3)
)
count_per_year("Plot", df)

}
