% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_data.R
\name{create_map}
\alias{create_map}
\title{Create a Custom Mapping File}
\usage{
create_map(dir = NULL)
}
\arguments{
\item{dir}{Directory to create the custom mapping file in.
Defaults to \code{~/.csmTools}.}
}
\value{
Path to the custom mapping JSON file.
}
\description{
Copies the core mapping file to a user directory for customization.
}
