% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_data.R
\name{download_dwd}
\alias{download_dwd}
\title{Download DWD past weather in a specific year for a given set of variables}
\usage{
download_dwd(
  vars = c("air_temperature", "precipitation", "solar_radiation", "dewpoint",
    "relative_humidity", "wind_speed"),
  year,
  stations,
  dir = tempdir()
)
}
\arguments{
\item{vars}{a character vector specifying the variable(s) to download}

\item{year}{an integer indicating the target year}

\item{stations}{a data frame listing target weather stations, a generated by the functions nearbyStations",
"nearbyStations_solar", and "find_stations"}
}
\value{
a list of named data frames, each containing weather data for one of the target variable(s)
}
\description{
Download DWD past weather data for crop modeling relevant variables: air temperature, precipitation, solar radiation,
dew point temperature, relative humidity, and wind speed. The function picks the best quality data based on the provided
list of weather stations separately for each target variable. Additional arguments for picking data types, like
temporal resolution, are specified within the weather station data frame.
}
