% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_soil_data.R
\name{download_sg}
\alias{download_sg}
\title{Download Soil Grids interpolation soil data for a given set of coordinates}
\usage{
download_sg(metadata = NULL, lat = NULL, lon = NULL, src = "isric")
}
\arguments{
\item{lat}{\subsection{}{
}}

\item{lon}{\subsection{}{
}}

\item{src}{\subsection{}{
}}

\item{vars}{a character vector specifying the variable(s) to download}

\item{id}{\subsection{}{
}}
}
\value{
\subsection{}{
}
}
\description{
\subsection{}{
}
}
