% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_phenology.R
\name{estimate_phenology}
\alias{estimate_phenology}
\title{Estimate important phenological dates based on crop and weather data whenever they are missing from the input dataset}
\usage{
estimate_phenology(sdata, wdata, crop, lat, lon, year, irrigated = FALSE)
}
\arguments{
\item{sdata}{a dataframe containing the target summary observed data, as generated by reshape_exp_data}

\item{wdata}{a dataframe containing the weather data for the target year, as generated by get_weather}

\item{crop}{a string indicating the target crop \link{currently as ICASA standard coding, e.g. WHT, MAZ, etc.}}

\item{lat}{a numeric indicating the latitude of the target experimental field}

\item{lon}{a numeric indicating the longitude of the target experimental field}

\item{year}{a numeric indicating the target year}

\item{irrigated}{a logical indicating whether the target field was irrigated or not (i.e., rainfed)}
}
\value{
the summary observed data with estimated phenological dates
}
\description{
Maturity dates are estimated based on the approaches from Waha et al. (2012) and Minoli et al. (2019) implemented
in the package cropCalendars (https://github.com/AgMIP-GGCMI/cropCalendars/tree/master)
Currently only handles maize, rice, sorghum, soybean, and wheat.
Anthesis date are currently estimated relative to maturity dates, using expert sources (temporary implementation).
Wheat: https://www.arvalis.fr/infos-techniques/stades-des-cereales-bientot-lepiaison
Maize: average difference from DSSAT default datasets (temperate)
}
