% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_management.R
\name{fetch_template}
\alias{fetch_template}
\title{Ensure User Has the Latest ICASA VBA Template}
\usage{
fetch_template(force = FALSE)
}
\arguments{
\item{force}{Logical. If \code{TRUE}, forces an update of the user template even if an older version exists, backing up the previous file. Default is \code{FALSE}.}
}
\value{
Invisibly returns a list with elements:
\describe{
\item{path}{The normalized path to the user template.}
\item{created}{Logical, \code{TRUE} if the template was newly created.}
\item{updated}{Logical, \code{TRUE} if the template was updated.}
}
}
\description{
Copies the latest version of the ICASA VBA template to the user's directory, updating and backing up as needed.
}
\details{
This function checks for the presence of the user template at \code{~/.csmTools/template_icasa_vba.xlsm}. If it does not exist, it is copied from the package's \code{extdata} directory.
If the user template exists, its version is compared to the package template using \code{check_template_version()}. If the package template is newer, the user is prompted to update.
If \code{force = TRUE}, the user template is backed up and replaced with the latest version from the package.
}
\examples{
# Ensure the user has the latest template (will prompt if update is available)
fetch_template()

# Force update and backup the old template
fetch_template(force = TRUE)

}
