% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_exp_data.R
\name{find_common_prefix}
\alias{find_common_prefix}
\title{Find the longest common prefix in a character vector}
\usage{
find_common_prefix(strings)
}
\arguments{
\item{strings}{A character vector of strings to compare.}
}
\value{
A character string representing the longest common prefix among the input strings.
}
\description{
This function takes a character vector and returns the longest common prefix shared by all elements.
}
\examples{
find_common_prefix(c("apple", "apricot", "ape")) # returns "ap"
find_common_prefix(c("dog", "cat", "mouse")) # returns ""

}
