% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_exp_design.R
\name{find_plot_col}
\alias{find_plot_col}
\title{Find Plot-Like Columns in a Data Frame}
\usage{
find_plot_col(df, year_col_name)
}
\arguments{
\item{df}{A data frame to search for plot-like columns.}

\item{year_col_name}{The name of the year column (as a string).}
}
\value{
A character vector of column names in \code{df} that are plot-like.
}
\description{
Identifies columns in a data frame that are "plot-like" according to the \code{\link{is_plot_like}} function.
A plot-like column is one that uniquely identifies plots within years, is repeated across years, and has the highest number of distinct values per year (see \code{is_plot_like} for details).
}
\examples{
# Example usage (assuming is_plot_like and get_pkeys are defined):
# df <- data.frame(Plot = c(1,1,2,2), Year = c(2000,2001,2000,2001), Value = c(5,6,7,8))
# find_plot_col(df, "Year")

}
\seealso{
\code{\link{is_plot_like}}
}
