% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_data.R
\name{find_stations}
\alias{find_stations}
\title{Select closest DWD weather stations}
\usage{
find_stations(lat, lon, min_date, params)
}
\arguments{
\item{lat}{a numeric depicting the latitude (y component) of the target location \link{degrees N/S, range 47:55}}

\item{lon}{a numeric depicting the longitude (x component) of the target location \link{degrees N/S, range 47:55}}

\item{min_date}{a character setting the maximum data recording starting date and minimum ending date}

\item{params}{\subsection{}{
}}
}
\value{
a data frame containing information on weather stations that best  meet the specified location, date range,
variables, and temporal resolution parameters
}
\description{
Download DWD past weather data for crop modeling relevant variables: air temperature, precipitation, solar radiation,
dew point temperature, relative humidity, and wind speed. The function picks the best quality data based on the provided
list of weather stations separately for each target variable. Additional arguments for picking data types, like
temporal resolution, are specified within the weather station data frame.
}
