% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_transform_template.R
\name{format_envmod_tbl}
\alias{format_envmod_tbl}
\title{Format the ICASA Environmental Modification Levels Table in a List}
\usage{
format_envmod_tbl(ls)
}
\arguments{
\item{ls}{A named list of data frames, typically representing sections of an experimental template, including "ENVIRON_MODIF_LEVELS".}
}
\value{
The input list with the "ENVIRON_MODIF_LEVELS" table formatted in wide format and renamed columns.
}
\description{
Transforms the "ENVIRON_MODIF_LEVELS" data frame in a list to a wide format, renaming columns for ICASA compatibility.
}
\details{
If the list contains a data frame named "ENVIRON_MODIF_LEVELS", this function:
\itemize{
\item Removes the \code{environ_parameter_unit} column.
\item Pivots the table to a wide format, with columns for each environmental parameter's modification code and value.
\item Renames columns to use the \code{environ_modif_code} and \code{environ_modif} prefixes for ICASA compatibility.
}
The transformed data frame replaces the original in the input list.

The function uses the \strong{dplyr} and \strong{tidyr} packages for data manipulation.
}
\examples{
\dontrun{
ls <- format_envmod_tbl(ls)
}

}
