% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_dataset.R
\name{format_icasa_headers}
\alias{format_icasa_headers}
\title{Format ICASA Column Headers}
\usage{
format_icasa_headers(
  dataset,
  header_type = "short",
  config_path = "./inst/extdata/datamodels.yaml"
)
}
\arguments{
\item{dataset}{The ICASA dataset as a named list of data frames.}

\item{header_type}{The target header style, either \code{"short"} or \code{"long"}.}

\item{config_path}{The path to the \code{datamodels.yaml} file.}
}
\value{
The dataset with renamed headers.
}
\description{
Swaps column headers between ICASA long 'variable name' format and short 'display code' format.
}
\details{
This utility function fetches the official ICASA dictionary from the URL specified in the configuration.
It then builds a lookup table to rename columns in the dataset to the desired \code{header_type}.
}
