% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_data.R
\name{format_weather}
\alias{format_weather}
\title{Format DWD weather data into a specified data standard}
\usage{
format_weather(data, lookup)
}
\arguments{
\item{data}{a character vector specifying the variable(s) to download}

\item{lookup}{a map for a target standards (TODO: replace by standard name, fetch the map internally)}
}
\value{
a data frame; weather data tabes formatted into the DSSAT standard structure
}
\description{
Transform DWD weather data into a specified data standard (as of 14.12.23: ICASA or DSSAT)
}
