% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dssat_file_templates.R
\docType{data}
\name{GENERAL_template}
\alias{GENERAL_template}
\title{DSSAT table templates}
\format{
An object of class \code{data.frame} with 1 rows and 14 columns.
}
\usage{
GENERAL_template
}
\description{
DSSAT table templates
}
\keyword{datasets}
