% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_metadata.R
\name{generate_dssat_id}
\alias{generate_dssat_id}
\title{Generate a Standard DSSAT Identifier}
\usage{
generate_dssat_id(type, institution, site = NA, year = NA, sequence_no = NA)
}
\arguments{
\item{type}{The type of entity code to generate.}

\item{institution}{The name of the institution (used for a 2-letter abbreviation).}

\item{site}{The name of the site (used for a 2-letter abbreviation).}

\item{year}{The year of the experiment (used for a 2-digit code).}

\item{sequence_no}{A numeric sequence or ID to ensure uniqueness.}
}
\value{
An 8-character, uppercase DSSAT identifier string.
}
\description{
Creates a standardized, 8-character DSSAT-compliant code for an experiment, field, cultivar, soil profile,
or weather station.
}
