% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_utils.R
\name{get_df_name}
\alias{get_df_name}
\title{Get the Name of a Data Frame Within a List}
\usage{
get_df_name(ls, df)
}
\arguments{
\item{ls}{A named list of data frames (or other objects).}

\item{df}{A data frame to search for within the list.}
}
\value{
A character vector of the name(s) of the list element(s) that are identical to \code{df}. Returns an empty character vector if no match is found.
}
\description{
Returns the name(s) of the element(s) in a list that are identical to a given data frame.
}
\details{
The function compares each element of the list \code{ls} to the provided data frame \code{df} using \code{identical}. It returns the name(s) of the list element(s) that are exactly identical to \code{df}.

This is useful for retrieving the name(s) of a data frame within a list of data frames, especially after subsetting or manipulation.
}
\examples{
df1 <- data.frame(a = 1:2)
df2 <- data.frame(a = 3:4)
mylist <- list(first = df1, second = df2)
get_df_name(mylist, df2)
# Returns \"second\"

}
