% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_data_relationships.R
\name{get_parent}
\alias{get_parent}
\title{Retrieve the parent table(s) of a table}
\usage{
get_parent(tbl, tbl_list)
}
\arguments{
\item{tbl}{a data frame}

\item{tbl_list}{a list of data frames}
}
\value{
the names of the parent tables of the focal table
}
\description{
Retrieve the parent table(s) of a table
}
