% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_soil_data.R
\name{get_soilGrids_dataverse}
\alias{get_soilGrids_dataverse}
\title{Download and Extract SoilGrids Profiles from Han et al. (2015); Harvard Dataverse}
\usage{
get_soilGrids_dataverse(dir = tempdir())
}
\arguments{
\item{dir}{Character. Directory in which to store and extract the SoilGrids profiles. Defaults to a temporary directory.}
}
\value{
The path to the directory containing the extracted SoilGrids profiles.
}
\description{
Downloads and extracts the SoilGrids by-country soil profile dataset
from the Harvard Dataverse (Han et al. 2015; https://doi.org/10.7910/DVN/1PEEY0), if not already present.
}
\details{
The function checks for the presence of the SoilGrids by-country zip file and its extracted directory in the specified location. If not found, it downloads the file from the Harvard Dataverse using the dataset DOI, saves it, and extracts its contents. If the data is already present, it simply returns the path to the extracted directory.

The function uses \code{dataset_files} and \code{get_file} to interact with the Dataverse API.
}
\examples{
\dontrun{
soil_dir <- get_soilGrids_dataverse()
list.files(soil_dir)
}

}
