% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_exp_design.R
\name{get_tbl}
\alias{get_tbl}
\title{Find a table based on its primary key}
\usage{
get_tbl(db, pkey)
}
\arguments{
\item{db}{a list of linked data frames}

\item{pkey}{a character, the name of the focal primary key}
}
\value{
a data frame with its name in the input list as an attribute
}
\description{
Find a table based on its primary key
}
