% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_management.R
\name{get_template_version}
\alias{get_template_version}
\title{Retrieve Template Version from Excel File}
\usage{
get_template_version(path)
}
\arguments{
\item{path}{Character string. The file path to the Excel workbook from which to retrieve the template version.}
}
\value{
A string version identifier.
}
\description{
Loads the Excel template workbook and extracts the version number.
}
\details{
This function uses \code{wb_load} to load the workbook and \code{wb_read} to read the value in a specific location in the workbook.
It assumes that the version information is always stored in this specific location.
}
\examples{
# Example usage:
# version <- get_template_version("path/to/template.xlsm")

}
