% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_exp_design.R
\name{get_treatments_col}
\alias{get_treatments_col}
\title{Find treatments column in a design data frame based on the years and plots column}
\usage{
get_treatments_col(design_tbl, years_col, plots_col)
}
\arguments{
\item{years_col}{the name of the years column}

\item{plots_col}{the name of the plots column}

\item{df}{a data frame}
}
\value{
a list containing the name of the treatment column and the number of replicates
}
\description{
Find treatments column in a design data frame based on the years and plots column
}
