% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_data.R
\name{get_weather_data}
\alias{get_weather_data}
\title{Wrapping function for downloading, formatting and imputing daily weather time series}
\usage{
get_weather_data(lon, lat, from, to, pars, res, src, output_path = NULL)
}
\arguments{
\item{lon}{(numeric) Longitude (x coordinate) of the target location \link{decimal degrees E/W, range 6:15}}

\item{lat}{(numeric) Latitude (y coordinate) of the target location \link{decimal degrees N/S, range 47:55}}

\item{from}{(character or date) Start date of the desired time range as YYYY-MM-DD (inclusive).}

\item{to}{(character or date) End date of the desired time range as YYYY-MM-DD (inclusive).}

\item{pars}{(character) A character string of length >= 1 specifying the target variables among:
\itemize{
\item \code{"air_temperature"} (default)
\item \code{"precipitation"}
\item \code{"solar_radiation"}
\item \code{"dewpoint"}
\item \code{"relative_humidity"}
\item \code{"wind_speed"}
\item \code{"par"} (photosynthetically active radiation)
\item \code{"evaporation"}
}}

\item{res}{(character vector) Temporal resolution for each weather variable.Must match length of \code{pars}. Supported values:
\itemize{
\item \code{"daily"} (default)
\item \code{"hourly"}
}}

\item{src}{(character) Source repository (currently only "nasa_power" is supported).}

\item{output_path}{(character) Optional file path to save the output.}
}
\value{
a list of dataframes containing daily weather for the requested location, time frame and variables.
}
\description{
Source repositories are currently limited to DWD Open Data Server, and imputation method to linear interpolation
}
