% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{growth_stages}
\alias{growth_stages}
\title{Growth stages reference data}
\format{
A data frame with 132 rows and 5 variables:
\describe{
\item{scale}{(character) The growth stage scale (bbch, zadoks)}
\item{growth_stage}{(character) Harmonized name of growth stage}
\item{growth_stage_scalespec}{(character) Scale-specific growth stage name}
\item{code}{(numeric) Numeric growth stage code}
\item{description}{(character) description of the growth stage}
}
}
\source{
Internal project data
}
\usage{
growth_stages
}
\description{
A dataset containing standardized growth stage scales for cereal crops.
Includes BBCH and Zadoks scales with corresponding codes.
}
\keyword{datasets}
