% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_data_relationships.R
\name{has_link}
\alias{has_link}
\title{Determine whether two data frames are linked by a common column}
\usage{
has_link(df1, df2, subset = NULL)
}
\arguments{
\item{df1}{the first data frame}

\item{df2}{the second data frame}

\item{subset}{character vector containing the column names that should be considered in the evaluation}
}
\value{
a logical stating whether the two data frames are dependent or not
}
\description{
Determine whether two data frames are linked by a common column
}
