% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_geoclusters.R
\name{haversine_dist}
\alias{haversine_dist}
\title{Calculate Haversine Distance}
\usage{
haversine_dist(lat1, lon1, lat2, lon2)
}
\arguments{
\item{lat1}{Latitude of the first point, in decimal degrees.}

\item{lon1}{Longitude of the first point, in decimal degrees.}

\item{lat2}{Latitude of the second point, in decimal degrees.}

\item{lon2}{Longitude of the second point, in decimal degrees.}
}
\value{
The distance between the two points, in meters.
}
\description{
Calculates the great-circle distance between two geographic points on Earth using the Haversine formula.
}
\examples{
# Calculate the distance between Paris and Lyon
haversine_dist(48.8566, 2.3522, 45.7640, 4.8357)

}
