% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_data.R
\name{headers_to_attr}
\alias{headers_to_attr}
\title{Convert weather metadata table into data frame attributes}
\usage{
headers_to_attr(df, header)
}
\arguments{
\item{df}{a data frame containing weather data}

\item{header}{a data frame containing weather metadata associated to df, as returned by}
}
\value{
a data frame containing weather data with attributes set to the values in the header table
}
\description{
Convert weather metadata table into data frame attributes
}
