% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_data.R
\name{impute_weather}
\alias{impute_weather}
\title{Impute missing values in daily weather time series data}
\usage{
impute_weather(df, na.rm = TRUE, rule = 2)
}
\arguments{
\item{df}{a data frame with weather time series data in which NAs are to be replaced}

\item{na.rm}{logical. If the result of the (spline) interpolation still results in leading and/or trailing NAs,
should these be removed (using na.trim)?}

\item{rule}{an integer (of length 1 or 2) specifying how leading and trailing NAs are handled.
If rule is 1 then NAs are returned for such points and if it is 2, the value at the closest data extreme is used.}
}
\value{
a data frame with NAs replaced according to the input arguments
}
\description{
So far only performs linear interpolation and extrapolation. Other methods will be added in the future.
(e.g., imputation with remote sensing NASA power data, sequential imputation, random forests, etc.)
}
