% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_utils.R
\name{intersect_dfs}
\alias{intersect_dfs}
\title{Find Common Data Frames Between Two Lists}
\usage{
intersect_dfs(list1, list2)
}
\arguments{
\item{list1}{A list of data frames.}

\item{list2}{A list of data frames to compare with \code{list1}.}
}
\value{
A list of data frames from \code{list1} that are also present in \code{list2}.
}
\description{
Returns the data frames from the first list that are also present (identical in content) in the second list.
}
\details{
The function compares data frames in \code{list1} and \code{list2} by converting each data frame to a string representation using \code{dput}. It then finds the intersection of these string representations and returns the data frames from \code{list1} that are present in both lists.

This approach allows for comparison of data frames by value, not by reference or name. Note that this method may be slow for very large data frames.
}
\examples{
df1 <- data.frame(a = 1:2)
df2 <- data.frame(a = 3:4)
df3 <- data.frame(a = 1:2)
list1 <- list(one = df1, two = df2)
list2 <- list(alpha = df3)
intersect_dfs(list1, list2)
# Returns a list containing df1, since df1 and df3 are identical

}
