% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_data_types.R
\name{is_treatment}
\alias{is_treatment}
\title{Identify if a management type is an experimental treatment}
\usage{
is_treatment(df, years_col, plots_col)
}
\arguments{
\item{df}{a data frame containing management data}

\item{years_col}{character string; name of the "year" column}

\item{plots_col}{character string; name of the plot identifier column}
}
\value{
a data frame containing the year column and a logical "is_trt"
}
\description{
Determines for each year whether a management type has different fixed value among plots within year (experimental treatment)
or is uniform among plots within year.
}
