% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_data_relationships.R
\name{is_unique_id}
\alias{is_unique_id}
\title{Identify the primary key(s) in a table}
\usage{
is_unique_id(df, col_name, type = c("int", "all"))
}
\arguments{
\item{df}{a data frame}

\item{col_name}{character string; name of the "year" column}

\item{type}{character string; specifies the data type for primary keys: "int" for integer, "all" for any type}
}
\value{
a logical stating whether the input variable is a unique identifier
}
\description{
Identify the primary key(s) in a table
}
