% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_metadata.R
\name{is_valid_dssat_code}
\alias{is_valid_dssat_code}
\title{Validate DSSAT Entity Codes}
\usage{
is_valid_dssat_code(x, item, framework = "dssat")
}
\arguments{
\item{x}{A character vector of codes to validate.}

\item{item}{The type of entity. One of "experiment", "cultivar", "field", "soil", or "weather_station".}

\item{framework}{The modeling framework (currently only "dssat" is supported).}
}
\value{
A logical vector indicating \code{TRUE} for each valid code.
}
\description{
Checks whether a character string matches the standard format for a given DSSAT entity code.
}
\details{
The function uses regular expressions to validate codes against DSSAT conventions:
\itemize{
\item \code{experiment}: institute (2L) site (2L) experiment year (2D) experiment number (2D) (e.g., "UFGA8201").
\item \code{cultivar}: institute (2L) cultivar number (4D) (e.g., "IB0015").
\item \code{field}: institute (2L) site (2L) field number (4D) (e.g., "UFGA0001").
\item \code{soil}: institute (2L) site (2L) epxeriment year (2D) soil profile number (4D)
\item \code{weather_station}: institute (2L) site (2L) (e.g., "UFGA").
}
}
