% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_data.R
\name{load_map}
\alias{load_map}
\title{Load a Mapping Data Frame}
\usage{
load_map(path = NULL)
}
\arguments{
\item{path}{Character. Path to the mapping JSON file.
If \code{NULL}, the default mapping file in the package is used.}
}
\value{
A data frame containing the mapping information.
}
\description{
Loads the core mapping data frame from a specified JSON file.
If no path is provided, a default mapping file included with the package is used.
}
