% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_data.R
\name{make_code_lookup}
\alias{make_code_lookup}
\title{Convert a mapping code list stored in a data frame as a string into a lookup table}
\usage{
make_code_lookup(vec)
}
\arguments{
\item{vec}{a code mapping list \link{format: list('source1: target1', 'source2: target2')} stored as a string}
}
\value{
a lookup dataframe with 2 columns: source (original data) and target (standard data)
}
\description{
The string format is analogous to the code_mapping field in standard ARDN SC2 json files around which the standard
mapping files for the package will be built. See: https://agmip.github.io/ARDN/Annotation_SC2.html
}
