% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_data.R
\name{map_codes}
\alias{map_codes}
\title{Map categorical data to standard codes}
\usage{
map_codes(df, map)
}
\arguments{
\item{df}{a data frame of the data to be mapped}

\item{map}{a lookup table of the categories to be madde to standard codes, as generated by 'made_code_lookup'}
}
\value{
a data frame with codes mapped to the format specified in the supplied map
}
\description{
Currently handles lookup tables as maps rather than json SC2 files, which will be implemented in the future.
}
