% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_data.R
\name{map_data}
\alias{map_data}
\title{Map a data table into a standard format}
\usage{
map_data(df, tbl_name, map, keep_unmapped = TRUE, col_exempt = NULL)
}
\arguments{
\item{df}{a data frame of the data to be mapped}

\item{map}{a lookup table detailing input and target headers, units, and codes}

\item{keep_unmapped}{a logical indicating whether to keep unmapped variables in the output}

\item{col_exempt}{a vector of column names to be kept into the output if keep_unmapped = FALSE}

\item{table_name}{the name of the focal crop experiment data section in the map}
}
\value{
a data frame with headers, units, and codes mapped to the format specified in the map
}
\description{
Currently only handles exact matches (header names, units, and codes). Future versions will handle more complex
data mapping (e.g., conditional, concatenation, etc.)
}
