% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_data.R
\name{map_headers}
\alias{map_headers}
\title{Map Column Headers to Standard Codes}
\usage{
map_headers(
  df,
  map,
  direction = c("to_icasa", "from_icasa"),
  header = c("short", "long")
)
}
\arguments{
\item{df}{A data frame containing the data to be mapped.}

\item{map}{A lookup table of categories to be mapped to standard codes,
as generated by \code{make_code_lookup}.}

\item{direction}{Character. Direction of the mapping:
\code{"to_icasa"} (to standard ICASA codes) or
\code{"from_icasa"} (from standard ICASA codes).}
}
\value{
A list with two elements:
\describe{
\item{data}{A data frame with headers mapped to the annotation format specified in \code{map_data}.}
\item{mapped}{A character vector of columns that were mapped.}
}
}
\description{
Maps the column headers of a data frame to standard codes using a lookup table.
}
