% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_data2.R
\name{merge_row_values}
\alias{merge_row_values}
\title{Aggregate a Vector to a Single Value}
\usage{
merge_row_values(x, sep = "; ")
}
\arguments{
\item{x}{A vector to consolidate.}

\item{sep}{A character string separator used for concatenation if multiple unique values are found. Defaults to \code{"; "}.}
}
\value{
A single value, which can be of the original data type (for a single unique value), a character string
(for multiple unique values), or \code{NA}.
}
\description{
Consolidates a vector into a single value. If only one unique non-NA value exists, it is returned as is.
if multiple unique values exist, they are pasted into a single character string.
}
\details{
This is a helper function often used inside \code{dplyr::summarise()} to aggregate character or factor columns.
The logic is as follows:
\enumerate{
\item All \code{NA} values are removed from the input vector \code{x}.
\item The unique values are identified.
\item If there is exactly one unique value, it is returned in its original data type.
\item If there are multiple unique values, they are collapsed into a single
string, separated by \code{sep}.
\item If the input contains no non-NA values, \code{NA} is returned.
}
}
