% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_data.R
\name{nearbyStations_solar}
\alias{nearbyStations_solar}
\title{Select DWD weather stations measuring solar radition closest to the given coordinates}
\usage{
nearbyStations_solar(lat, lon, res, max_radius = 50)
}
\arguments{
\item{lat}{a numeric depicting the latitude (y component) of the target location \link{degrees N/S, range 47:55}}

\item{lon}{a numeric depicting the longitude (x component) of the target location \link{degrees N/S, range 47:55}}

\item{res}{a character setting the maximum data recording starting date and minimum ending date}

\item{per}{\subsection{}{
}}
}
\value{
a data frame containing information on weather stations that best  meet the specified location, date range,
variables, and temporal resolution parameters
}
\description{
Select DWD stations within a given radius around a set of coordinates. Workaround for rdwd::nearbyStations not
functioning for solar raidation data
}
