% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_tbls.R
\name{normalize}
\alias{normalize}
\title{Normalize a character vector by trimming whitespace}
\usage{
normalize(x)
}
\arguments{
\item{x}{A vector to normalize.}
}
\value{
A character vector with whitespace trimmed.
}
\description{
Converts input to character and removes leading and trailing whitespace.
}
\examples{
normalize("  hello world  ")
normalize(c("  a", "b  ", "  c  "))

}
