% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_exp_data.R
\name{read_exp_data}
\alias{read_exp_data}
\title{Read and standardize the name of agricultural experiment tables (BonaRes LTE data model)}
\usage{
read_exp_data(dir, extension = "csv", origin = "bonares")
}
\arguments{
\item{dir}{Character. Path to the directory containing the data files.}

\item{extension}{Character. File extension to look for (default is "csv").}

\item{origin}{Character. Origin of the data (default is "bonares"). Placeholder; currently not used in the function.}
}
\value{
A named list of data frames, each corresponding to a file in the directory.
}
\description{
Reads all files with a specified extension from a directory into a list of data frames.
The function also simplifies the names of the resulting list elements by removing common prefixes and file extensions.
(based on BonaRes LTE file name conventions)
}
\details{
The function reads all files with the specified extension in the given directory using \code{read.csv} with UTF-8 encoding.
The names of the list elements are simplified by removing the common prefix, file extension, and (optionally) a table number pattern.
}
\examples{
\dontrun{
data_list <- read_exp_data("path/to/data", extension = "csv")
}

}
