% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_input.R
\name{read_json_dataset}
\alias{read_json_dataset}
\title{Reads a dataset from a JSON file created by \code{write_json_dataset},
reconstructing dataframes, tibbles, and preserving attributes recursively.}
\usage{
read_json_dataset(file_path)
}
\arguments{
\item{file}{A character string specifying the path to the input JSON file.}
}
\description{
Reads a dataset from a JSON file created by \code{write_json_dataset},
reconstructing dataframes, tibbles, and preserving attributes recursively.
}
