% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_utils.R
\name{remove_all_na_cols}
\alias{remove_all_na_cols}
\title{Remove columns containing only NAs from a data frame}
\usage{
remove_all_na_cols(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A data frame with columns containing only NAs removed.
}
\description{
Remove columns containing only NAs from a data frame
}
\examples{
df <- data.frame(a = c(NA, NA), b = c(1, NA), c = c(NA, NA))
remove_all_na_columns(df)

}
