% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_dssat.R
\name{reset_id}
\alias{reset_id}
\title{Reset index variables to start at 1}
\usage{
reset_id(df, id_col)
}
\arguments{
\item{df}{a data frame with rows identified by an integer index variable}

\item{id_col}{a length 1 character containing the name of the index column in df}
}
\value{
a data frame; the input data with a reset index variable
}
\description{
Includes a special case for ignoring 0s (ID in treatment matrix with no linked record in a management section)
}
