% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_exp_data.R
\name{reshape_exp_data}
\alias{reshape_exp_data}
\title{Transform BonaRes LTE dataset into the ICASA-format}
\usage{
reshape_exp_data(
  db,
  mother_xml,
  design_tbl_name,
  data_model = "bonares-lte_de",
  header_type = "long"
)
}
\arguments{
\item{db}{a list of data frames composing a crop experiment dataset.}

\item{mother_xml}{a path or URL to the metadata XML file of the mother table.}

\item{design_tbl_name}{a character string specifying the name of the data frame within \code{db} that serves as the mother table (experimental design).}

\item{data_model}{a character string specifying the data model of the input dataset (e.g., "bonares-lte_de").}
}
\value{
a list containing the reshaped crop experiment data in ICASA format.
}
\description{
This function identifies the components of a crop experiment data set, map terms into ICASA terminology
and re-arranges the table according to the ICASA data model
(see ICASA standards: https://github.com/DSSAT/ICASA-Dictionary).
}
\details{
The \code{reshape_exp_data} function performs a series of steps to transform a raw crop experiment dataset,
typically provided as a list of data frames, into a structure compliant with the ICASA standards.

\strong{Workflow:}
\itemize{
\item \strong{Metadata Extraction:} Reads metadata from a specified XML file (\code{mother_xml}) and appends it to the \code{db} list.
\item \strong{Experimental Design Identification:} Uses the \code{design_tbl_name} to identify the core components
of the experimental design, such as years, plots, and treatments. It also identifies primary and foreign keys and data attributes within the dataset.
\item \strong{Data Merging and Cleaning:} Recursively merges data frames in \code{db} based on identified keys
and removes columns that contain only \code{NA} values.
\item \strong{Date Formatting:} Standardizes all date columns across the dataset to the 'yyyy-mm-dd' format.
\item \strong{ICASA Mapping:} Maps column names within each data frame to their corresponding ICASA headers
based on the specified \code{data_model}. This step uses \code{load_map()} and \code{map_data()}.
\item \strong{Dataset Projection:} Further structures the data into distinct ICASA sections (e.g., 'PLOT_DETAILS', 'METADATA',
'SUMMARY', 'TIME_SERIES', 'WEATHER_METADATA', 'WEATHER_DAILY', 'SOIL_METADATA', 'SOIL_LAYERS', and various 'MANAGEMENT' sections) using \code{project_dataset()}.
\item \strong{Management Data Formatting:} A crucial step is the reformatting of management data.
This involves identifying unique event regimes within each experimental year and assigning unique IDs. It also creates a "TREATMENTS" matrix that links experimental years, plots, treatments, and their associated management regimes.
}

The function aims to automate the structural transformation, assuming that variable content is consistent with the chosen \code{data_model}. Users may need to perform additional variable mapping or data cleaning outside this function if their data deviates significantly from the expected structure or if more granular control over variable transformation is required.
}
