% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_transform_template.R
\name{reshape_to_model_dssat}
\alias{reshape_to_model_dssat}
\title{TEMP reshape function from icasa to dssat approach;
later integrate to a generic melt-explode-map sequence}
\usage{
reshape_to_model_dssat(
  vector,
  map_path,
  input_model = "icasa",
  output_model = "dssat"
)
}
\description{
TEMP reshape function from icasa to dssat approach;
later integrate to a generic melt-explode-map sequence
}
