% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_transform_template.R
\name{split_experiments}
\alias{split_experiments}
\title{Split a List of Data Frames into Experiments with Associated Soil and Weather Data}
\usage{
split_experiments(ls, keep_empty = FALSE, keep_na_cols = FALSE)
}
\arguments{
\item{ls}{A named list of data frames, typically representing sections of an experimental template, including experiment, soil, and weather data.}
}
\value{
A list of lists, each representing an experiment with its associated data frames.
}
\description{
Splits a list of data frames into separate experiments, attaching relevant soil and weather data to each experiment based on identifier columns.
}
\details{
The function determines the appropriate identifier columns for experiments, soil, and weather data based on the column names in the first data frame. It splits experiment data frames by experiment ID, then attaches soil and weather data frames that match the relevant profile or station IDs for each experiment. All-NA columns in measured data sections are removed.

This is useful for organizing a combined dataset into a list of experiments, each with its associated environmental data, for further analysis or export.

The function uses the \strong{dplyr} package for data manipulation and relies on the helper function \code{revert_list_str}.
}
\examples{
\dontrun{
experiments <- split_experiments(ls)
}

}
