% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_metadata.R
\name{structure_dssat_mngt}
\alias{structure_dssat_mngt}
\title{Structure DSSAT Management and Metadata}
\usage{
structure_dssat_mngt(dataset)
}
\arguments{
\item{dataset}{The mapped dataset as a list of data frames.}
}
\value{
The dataset with standardized \code{GENERAL}, \code{FIELDS}, and \code{CULTIVARS} tables.
}
\description{
Processes and standardizes the core metadata and management metadata of a DSSAT dataset
}
\details{
This function performs several key operations:
\enumerate{
\item \strong{Enriches Field Data:} It uses external APIs (\code{elevatr}, \code{tidygeocoder}) to fetch and fill missing elevation
and location details (city, region, country).
\item \strong{Generates Standard IDs:} It calls \code{generate_dssat_id()} to create compliant identifiers for
experiments (\code{EXP_ID}), fields (\code{ID_FIELD}), and cultivars (\code{INGENO}) if they are missing or invalid.
\item \strong{Links Data:} It ensures that the generated IDs are consistent across the
\code{GENERAL}, \code{FIELDS}, and \code{CULTIVARS} tables.
}
}
