% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_metadata.R
\name{structure_dssat_sol}
\alias{structure_dssat_sol}
\title{Structure DSSAT Soil Data}
\usage{
structure_dssat_sol(dataset)
}
\arguments{
\item{dataset}{The mapped dataset as a list of data frames.}
}
\value{
The dataset with soil tables replaced by standardized \code{SOIL} and \code{SOIL_HEADER} data frames.
}
\description{
Consolidates, enriches, and standardizes soil profile data for a DSSAT dataset.
}
\details{
This function merges the \code{SOIL_META}, \code{SOIL_GENERAL}, and \code{SOIL_LAYERS} tables into a single, comprehensive soil
data frame compatible with file writing functions from the 'DSSAT' package.
It then enriches this data by:
\enumerate{
\item Fetching missing location metadata using external geocoding APIs.
\item Generating a standardized DSSAT soil ID (\code{PEDON}) if the existing one is invalid and updating the linked.
\code{FIELDS} table accordingly
}
}
