% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_metadata.R
\name{structure_icasa_mngt}
\alias{structure_icasa_mngt}
\title{Normalize ICASA Management Data into Regimes}
\usage{
structure_icasa_mngt(ls, master_key, year_col, treatment_col, plot_col)
}
\arguments{
\item{ls}{A named list of data frames containing the ICASA management sections
(e.g., \code{TREATMENTS}, \code{FERTILIZERS}, \code{PLANTINGS}).}

\item{master_key}{The column name of the master identifier (e.g., "experiment_ID").}

\item{year_col}{The column name for the experimental year.}

\item{treatment_col}{The column name for the treatment number.}

\item{plot_col}{The column name for the plot identifier.}
}
\value{
A list containing the \code{management} list (formatted event tables) and  the \code{management_matrix}
(the updated \code{TREATMENTS} table).
}
\description{
Restructures a list of ICASA management tables by aggregating unique sequences of events into standardized
"management regimes."
}
\details{
The core purpose of this function is to simplify complex, event-based management data. It processes each management
table (e.g., \code{FERTILIZERS}, \code{TILLAGE}) and identifies unique schedules of operations applied across plots or fields.
These unique schedules, or "regimes," are then assigned a new, consistent integer ID.
The function orders these new regime IDs based on the original \code{treatment_number} to maintain a logical sequence.

The final output is a list containing two main components:
\enumerate{
\item \strong{\code{management}}: A list of the formatted management tables, now containing the new regime IDs.
\item \strong{\code{management_matrix}}: The updated \code{TREATMENTS} table, which serves as a  cross-reference linking each
treatment to its corresponding regime IDs for all management types.
}
}
