% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_metadata.R
\name{structure_metadata}
\alias{structure_metadata}
\title{Format metadata documentation for ICASA/DSSAT datasets into the target data model}
\usage{
structure_metadata(dataset, data_model = c("icasa", "dssat"))
}
\arguments{
\item{dataset}{An ICASA/DSSAT dataset as a list of named data frames with unstructured metadata}

\item{data_model}{the target data model (ICASA or DSSAT)}

\item{section}{!deprecated, left from old version}
}
\value{
An ICASA/DSSAT dataset as a list of named data frames with metadata structured
to the target data model
}
\description{
This function documents metadata (provenance, spatial and temporal coverage) and naming
conventions in compliance with a speicifed (meta)data model
}
