% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_data.R
\name{transform_data}
\alias{transform_data}
\title{Apply Wrangling Functions from a Map to a Data Frame}
\usage{
transform_data(df, map, ...)
}
\arguments{
\item{df}{A data frame to be transformed.}

\item{map}{A data frame containing the mapping information and functions
(full map available with \code{load_map()}).}

\item{...}{Additional arguments (currently unused).}
}
\value{
A data frame with transformations applied as specified in the map.
}
\description{
Applies a set of wrangling functions, as specified in a mapping data frame,
to the columns of a data frame.
}
