% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_data2.R
\name{update_output_data}
\alias{update_output_data}
\title{Update a Section in a List with New Data}
\usage{
update_output_data(output_data, target_section, data_to_add)
}
\arguments{
\item{output_data}{A list object where each element is typically a data frame.}

\item{target_section}{A character string; the name of the list element to add or update.}

\item{data_to_add}{A data frame containing the data to be added or merged.}
}
\value{
The \code{output_data} list with the specified section updated.
}
\description{
Adds or merges a data frame into a specific element of a list. The function intelligently decides whether to perform
a full join or a column-bind.
}
\details{
This function updates a list (\code{output_data}) by adding or modifying the data frame within a \code{target_section}.
\itemize{
\item If the \code{target_section} does not exist, it is created.
\item If the section exists, the function identifies common columns between the existing and new data.
\itemize{
\item A \code{dplyr::full_join()} is performed using these common columns as keys.
\item If no common columns exist, a \code{dplyr::bind_cols()} is performed instead.
}
\item It specifically suppresses the default \code{dplyr} warning for many-to-many relationships and issues a more informative, context-aware warning.
}
}
