% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_output.R
\name{write_json_dataset}
\alias{write_json_dataset}
\title{Writes a list of dataframes (dataset) as a unique JSON file,
preserving attributes and nested structures, recursively.}
\usage{
write_json_dataset(dataset, file, pretty = TRUE, auto_unbox = TRUE, ...)
}
\arguments{
\item{dataset}{An R list where elements can be dataframes (potentially with custom attributes),
other lists (potentially nested dataframes), or NULL.}

\item{file}{A character string specifying the path to the output JSON file.}

\item{pretty}{Logical. If TRUE (default), the JSON will be pretty-printed.}

\item{auto_unbox}{Logical. Passed to jsonlite::toJSON. If TRUE, single-element arrays are unboxed.
Set to TRUE to prevent attributes like single character strings from becoming arrays.}

\item{...}{Additional arguments passed to jsonlite::write_json.}
}
\description{
Writes a list of dataframes (dataset) as a unique JSON file,
preserving attributes and nested structures, recursively.
}
