% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_dssat.R
\name{write_wth2}
\alias{write_wth2}
\title{Write multiple types of standard DSSAT files before}
\usage{
write_wth2(wth, file_name)
}
\arguments{
\item{ls}{a list DSSAT-formated crop experiment data}

\item{path}{a character vector storing the path to the directory where the files will be written}
}
\description{
Only handles input tables that have been previously mapped to the DSSAT standard format
}
