% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_dataset.R
\name{apply_mappings}
\alias{apply_mappings}
\title{Apply a Full Set of Mapping Rules to a Dataset}
\usage{
apply_mappings(input_data, input_model, output_model, config, ...)
}
\arguments{
\item{input_data}{The source dataset as a list of data frames.}

\item{input_model}{The character name of the source model.}

\item{output_model}{The character name of the target model.}

\item{config}{The parsed \code{datamodels.yaml} configuration object.}

\item{...}{Optional arguments to be passed down to \code{process_actions}.}
}
\value{
The final, mapped dataset as a list of data frames.
}
\description{
Apply a set of rules set in a YAML file to convert a dataset from a source to a target model.
}
\details{
This function manages the high-level transformation workflow. It first prepares the data by propagating the master key.
It then runs in two phases:
\enumerate{
\item \strong{Phase 1 (Transformation):} Executes all data generation and manipulation rules (e.g., lookups, calculations)
via \code{process_actions()}.
\item \strong{Phase 2 (Formatting):} Applies all formatting rules (e.g., dates, numbers) to the newly created tables.
}
}
