% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_dssat.R
\name{build_filea}
\alias{build_filea}
\title{Format observed summary data as a DSSAT input table}
\usage{
build_filea(ls, title = NULL, site_code = NA_character_)
}
\arguments{
\item{ls}{a list of data frames containing crop experiment data ### (File A)}

\item{title}{a length 1 character vector specifying the title of the dataset}

\item{site_code}{a length 1 character vector provide the DSSAT standard code for the experimental site}
}
\value{
a data frame with codes mapped to the format specified in the supplied map
}
\description{
Format observed summary data as a DSSAT input table
}
