% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_dssat.R
\name{build_filex}
\alias{build_filex}
\title{Format crop management data as DSSAT input tables}
\usage{
build_filex(ls, title = NULL, site_code = NA_character_)
}
\arguments{
\item{ls}{a list of data frames containing crop management data, to be formatted into DSSAT input tables}

\item{title}{a length 1 character vector specifying the title of the dataset}

\item{site_code}{a length 1 character vector provide the DSSAT standard code for the experimental site}
}
\value{
a data frame with codes mapped to the format specified in the supplied map
}
\description{
Format crop management data as DSSAT input tables
}
