% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_dssat.R
\name{build_sol}
\alias{build_sol}
\title{Format observed time series data as a DSSAT input table}
\usage{
build_sol(ls)
}
\arguments{
\item{ls}{a list of data frames containing soil profile data (layers + surface as separate data frames)}
}
\value{
a data frame containing soil profile data formatted as a DSSAT input
}
\description{
Format observed time series data as a DSSAT input table
}
