% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_dssat.R
\name{build_wth}
\alias{build_wth}
\title{Format observed time series data as a DSSAT input table}
\usage{
build_wth(ls)
}
\arguments{
\item{ls}{a list of data frames containing weather data (daily + header as separate data frames) and station
metadata as comments}
}
\value{
a data frame containing soil profile data formatted as a DSSAT input
}
\description{
Currently contains a workaround to correct date formatting issues with the write_wth function that leads to
failed simulation
}
