% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_dssat.R
\name{collapse_cols}
\alias{collapse_cols}
\title{Nest selected columns into single cells}
\usage{
collapse_cols(df, cols)
}
\arguments{
\item{df}{a data frame to collapse}

\item{cols}{a character vector of column names to collapse}
}
\value{
a data frame where the specified rows have been collapsed into single rows based on all unique combinations of the remaining columns
}
\description{
This function is necessary to format composite management tables, such as INITIAL_CONDITIONS and IRRIGATION
}
