% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_data.R
\name{convert_unit}
\alias{convert_unit}
\title{Convert Numeric Vector Between Units}
\usage{
convert_unit(x, u1, u2)
}
\arguments{
\item{x}{Numeric vector to be converted.}

\item{u1}{Input unit (as a string) in which \code{x} is expressed.}

\item{u2}{Target unit (as a string) for the output.}
}
\value{
Numeric vector converted to the target unit.
}
\description{
Converts a numeric vector from one unit to another, using unit strings.
}
