% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_phenology.R
\name{estimate_phenology}
\alias{estimate_phenology}
\title{Estimate important phenological dates for model calibration based on existing data}
\usage{
estimate_phenology(
  dataset,
  data_model = "dssat",
  method = "crop parameters",
  wheat_season = NULL
)
}
\arguments{
\item{dataset}{a list containing an input DSSAT dataset as generated by 'compile_model_dataset'}

\item{data_model}{a character specifying the output data model (for now only DSSAT is handled)}

\item{method}{a character specifying the metod for phenology estimation ("crop parameters" or "crop calendars")}

\item{wheat_season}{an optional argument specifying the seasionality of wheat ("winter" or "spring")}
}
\value{
a list of DSSAT input data components with estimated phenology appended to the observed data (file A)
}
\description{
Maturity dates are estimated based on the approaches from Waha et al. (2012) and Minoli et al. (2019) implemented
in the package cropCalendars (https://github.com/AgMIP-GGCMI/cropCalendars/tree/master)
Currently only handles maize, rice, sorghum, soybean, and wheat.
}
