% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_dssat.R
\name{format_table}
\alias{format_table}
\title{Prepare data tables for export as various DSSAT input file types}
\usage{
format_table(df, template)
}
\arguments{
\item{df}{a data frame of the data to be formatted}

\item{template}{a model data frame for the focal file type, as generated in 'dssat_file_templates.R'}
}
\value{
a data frame; a formatted DSSAT input table

lubridate as.POSIXct\ # is part of base now
}
\description{
Prepare data tables for export as various DSSAT input file types
}
