% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_utils.R
\name{get_df_name}
\alias{get_df_name}
\title{Extract the name of a data frame from a list of data frame}
\usage{
get_df_name(ls, df)
}
\arguments{
\item{ls}{a list of data frames containing the focal data frame}

\item{df}{a data frame; the focal table}
}
\value{
a length 1 character vector corresponding to the name of the mother table in the original dataset
}
\description{
Extract the name of a data frame from a list of data frame
}
