% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_utils.R
\name{is_date}
\alias{is_date}
\title{Determine whether a vector is date data}
\usage{
is_date(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a logical value indicating whether x is a date or not
}
\description{
Determine whether a vector is date data
}
