% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_dataset.R
\name{merge_row_values}
\alias{merge_row_values}
\title{Aggregate a Vector to a Single Value}
\usage{
merge_row_values(x, sep = "; ")
}
\arguments{
\item{x}{The vector to consolidate.}

\item{sep}{The separator for concatenation if multiple unique values are found.}
}
\value{
A single value (original type, string, or \code{NA}).
}
\description{
A helper function to consolidate a vector, typically used within \code{summarise()}.
It returns a single unique value or a concatenated string if multiple exist.
}
