% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_tbls.R
\name{merge_tbls}
\alias{merge_tbls}
\title{Merge data tables}
\usage{
merge_tbls(
  parent_list,
  child_list,
  type = c("parent-child", "child-parent", "bidirectional"),
  drop_keys = TRUE
)
}
\arguments{
\item{parent_list}{a list containing the parent data frames}

\item{child_list}{a list containing the child data frames}

\item{type}{type of join between parent and child tables; "parent-child" (parent primary key is the join column);
"child-parent" (child primary key is the join column); "bidirectional" (looks for both parent-child and child-parent joins)}

\item{drop_keys}{logical; keep join keys in merged tables}
}
\value{
a list containing the parent tables with their joined attributes from the child tables
}
\description{
Detects join keys between two related data tables and join tables accordingly
}
