% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_dataset.R
\name{process_actions}
\alias{process_actions}
\title{Process a Sequence of Data Transformation Actions}
\usage{
process_actions(actions, full_input_df, full_dataset, master_key_name, ...)
}
\arguments{
\item{actions}{A list of action definitions from a YAML rule.}

\item{full_input_df}{The initial data frame for the rule.}

\item{full_dataset}{The complete dataset, used for lookups.}

\item{master_key_name}{The name of the master key, used for default joins.}

\item{...}{Optional arguments passed down from \code{convert_dataset}.}
}
\value{
A single data frame representing the result of all applied actions.
}
\description{
The core mapping engine that executes a sequence of declarative actions defined in a YAML rule to transform a data frame.
}
\details{
This function iterates through a list of actions and dispatches the appropriate logic based on \code{action$type}.
Supported actions include:
\itemize{
\item \strong{\code{lookup_and_add}}: Joins data from another table.
\item \strong{\code{map_values}}: Translates values using a key-value map.
\item \strong{\code{rowwise_transform}}: Applies a formula to generate a new column.
\item \strong{\code{concatenate_columns}}: Pastes multiple columns into one.
\item \strong{\code{coalesce_columns}}: Finds the first non-NA value across a set of columns.
\item \strong{\code{rename_column} / \code{delete_column}}: Explicit. Basic column manipulation.
\item \strong{\code{filter_rows}}: Subsets the data based on column presence.
\item \strong{\code{split_and_map_text}}: Split column into multiple columns
\item \strong{\code{add_column}}: (Conditionally) add a single-value attribute (column)
\item \strong{\code{add_column_conditional}}: Add column and conditionally populate with set values
}
}
