% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata.R
\name{read_metadata}
\alias{read_metadata}
\title{Format metadata for use as input arguments in crop experiment data wrangling functions}
\usage{
read_metadata(path, repo, sheet = NULL)
}
\arguments{
\item{path}{a length 1 character vector containing the url to the metadata file}

\item{repo}{a length 1 character containing the name or acronym of the data repository from which the metadata was downloaded}

\item{sheet}{\subsection{}{
}}
}
\value{
a formatted metadata item (TODO: define standard schema) ####
}
\description{
Currently handles only old xml formats of the BonaRes repository (no longer available as of 10.01.2024)
}
