% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_exp_data.R
\name{reshape_exp_data}
\alias{reshape_exp_data}
\title{Reshape crop experiment data into a standard ICASA format}
\usage{
reshape_exp_data(db, metadata, mother_tbl)
}
\arguments{
\item{db}{a list of data frames composing a crop experiment dataset}

\item{metadata}{a metadata table described the dataset, as returned by read_metadata (BonaRes-LTFE format)}

\item{mother_tbl}{a data frame; the mother table of the dataset, i.e., describing the experimental design (years, plots_id, treatments_ids, replicates)}
}
\value{
a list containing the reshaped crop experiment data
}
\description{
This function identifies the components of a crop experiment data set and re-arrange them into ICASA sections
(see ICASA standards: https://docs.google.com/spreadsheets/u/0/d/1MYx1ukUsCAM1pcixbVQSu49NU-LfXg-Dtt-ncLBzGAM/pub?output=html)
Note that the function does not handle variable mapping but only re-arranges the data structure.
}
