% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_dataset.R
\name{resolve_column_name}
\alias{resolve_column_name}
\title{Find a Column Name Using Synonyms}
\usage{
resolve_column_name(name_spec, available_cols)
}
\arguments{
\item{name_spec}{A character string or a list with \verb{$header} and \verb{$synonyms} elements defining the names to search for.}

\item{available_cols}{A character vector of column names to search within.}
}
\value{
The character string of the first matching column name, or \code{NULL} if no match is found.
}
\description{
Finds a column in a list of available columns using a flexible specification that can include a primary header and
a list of synonyms.
}
