% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_utils.R
\name{safe_aggregate}
\alias{safe_aggregate}
\title{Skip NAs in aggregation functions regardless of whether na.rm is an argument}
\usage{
safe_aggregate(x, FUN, ...)
}
\arguments{
\item{x}{a function name}
}
\value{
a function name with na.rm argument as appropriate
}
\description{
Skip NAs in aggregation functions regardless of whether na.rm is an argument
}
