% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_utils.R
\name{substr_rows}
\alias{substr_rows}
\title{Subset rows from two data frames}
\usage{
substr_rows(df1, df2)
}
\arguments{
\item{df1}{the first data frame}

\item{df2}{the second data frame}
}
\value{
a data frame containing the rows that are unique to each data frame
}
\description{
Subsets rows from two data frames by returning te rows that are unique to each one
}
