% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_data_types.R
\name{tag_data_type}
\alias{tag_data_type}
\title{Identify the type of experimental data}
\usage{
tag_data_type(db, years_col, plots_col, plots_len, max_events = 8)
}
\arguments{
\item{db}{a list of data frames}

\item{years_col}{character string; name of the "year" column}

\item{plots_col}{character string; name of the plot identifier column}

\item{plots_len}{integer; length of the plot identifier (i0.e., number of plots)}

\item{max_events}{integer; maximum number of management event of the same type (e.g., irrigation) within one year; default is 8}
}
\value{
the input database tagged as "management", "observed" or "other" data category (as attr(., "category"))
}
\description{
Determines whether a data table contains management data (i.e., fixed experimental parameters), observed data
(i.e., parameters measured during the experiment) or other attributes (i.e., neither management nor observed data)
}
