% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_dataset.R
\name{update_output_data}
\alias{update_output_data}
\title{Update a Section in the Output Dataset}
\usage{
update_output_data(mapping_uid, output_data, target_section, data_to_add)
}
\arguments{
\item{mapping_uid}{The UID of the current mapping rule, used for context in warnings.}

\item{output_data}{The list object representing the entire output dataset.}

\item{target_section}{The character name of the list element (data frame) to update.}

\item{data_to_add}{The data frame to be added or merged.}
}
\value{
The \code{output_data} list with the target section updated.
}
\description{
Intelligently merges or adds a data frame to a target section within the main output list.
}
\details{
This function updates the main \code{output_data} list. If the \code{target_section} does not yet exist, it is created.
If it does exist, the function joins the new data (\code{data_to_add}) with the existing data. The join is performed on all
common columns. If no common columns exist, a column-bind (\code{bind_cols}) is performed instead. It also suppresses and
replaces the default \code{dplyr} many-to-many join warning with a more informative message.
}
